dj_mysql_docker/backend/requirements/prod.txt
asgiref==3.8.1
Django==4.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
mysqlclient==2.2.4
PyJWT==2.8.0
python-decouple==3.8
sqlparse==0.5.0
typing_extensions==4.12.2
